fixup! Added LMS support for OpenSSL commandline signature verification using pkeyutl.

+++ providers/implementations/encode_decode/lms_codecs.c
@@ -158,11 +158,11 @@ int ossl_lms_key_to_text(BIO *out, const LMS_KEY *key, int selection)
     }
     if (BIO_printf(out, "lms-type: %s-N%u-H%u (0x%x)\n",
             get_digest(lms_params->digestname),
-            lms_params->n, lms_params->h, lms_params->lms_type) <= 0)
+            lms_params->n, lms_params->h, (int)lms_params->lms_type) <= 0)
         return 0;
     if (BIO_printf(out, "lm-ots-type: %s-N%u-W%u (0x%x)\n",
             get_digest(ots_params->digestname),
-            ots_params->n, ots_params->w, ots_params->lm_ots_type) <= 0)
+            ots_params->n, ots_params->w, (int)ots_params->lm_ots_type) <= 0)
         return 0;
     if (!ossl_bio_print_labeled_buf(out, "Id:", key->Id, 16))
         return 0;
